requirements.txt
flask==2.0.1
numpy==1.21.0
opencv-python==4.5.3.56
pytesseract==0.3.8
pillow==8.3.1
pymupdf==1.18.14
pymongo==3.12.0
# ollama==0.1.0
google-generativeai==0.3.0 